refactor: 패키지 구조 수정, 코드 스타일 통일
