+++ src/main/java/racinggame/ConsoleView.java
@@ -9,7 +9,6 @@ import java.util.Scanner;
 import java.util.stream.Collectors;
 
 public final class ConsoleView {
-
     private static final Scanner scanner = new Scanner(System.in);
 
     public static void printCarNamesInputPrompt() {
@@ -48,5 +47,4 @@ public final class ConsoleView {
         racingCarGame.showStatus();
         System.out.println();
     }
-
 }
