+++ src/main/java/racinggame/domains/MoveChecker.java
@@ -1,7 +1,6 @@
 package racinggame.domains;
 
 public final class MoveChecker {
-
     private MoveChecker() {
         throw new AssertionError("정적 메서드만 보유한 객체입니다.");
     }
@@ -9,5 +8,4 @@ public final class MoveChecker {
     public static boolean check(int randomNumber) {
         return randomNumber >= 4;
     }
-
 }
