+++ src/test/java/calculator/ParserTest.java
@@ -1,5 +1,6 @@
 package calculator;
 
+import calculator.domains.Parser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
