+++ src/test/java/racinggame/domains/RacingCarGameTest.java
@@ -10,12 +10,10 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 
 @DisplayName("")
 public class RacingCarGameTest {
-
     @Test
     @DisplayName("게임 객체 생성 여부 확인")
     void isGameExist() {
         RacingCarGame g = RacingCarGame.newInstance();
         assertThatCode(() -> Objects.requireNonNull(g)).doesNotThrowAnyException();
     }
-
 }
